backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Redis
redis==5.0.1

# HTTP Requests
requests==2.31.0
aiohttp==3.9.1

# Data Processing
pandas==2.1.4
numpy==1.25.2

# Technical Analysis (install separately after C library)
# TA-Lib==0.4.28

# Environment Variables
python-dotenv==1.0.0

# Data Validation
pydantic==2.5.1

# Background Tasks (for later)
celery==5.3.4

# Development
pytest==7.4.3
python-multipart==0.0.6

# API Documentation
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Market Data Providers
yfinance==0.2.61
pandas-ta==0.3.14b0  # Technical analysis for pandas

python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

email-validator==2.1.0
